// object-storage/models/CreateFolder201Response.ts
/**
 * Object Storage API Documentaion
 * A scalable, reliable, and cost effective Object Storage solution to support your application from liara-cloud
 *
 * OpenAPI spec version: 1.0.0
 * Contact: support@liara.ir
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { CreateFolder201ResponseData } from '../models/CreateFolder201ResponseData';
import { HttpFile } from '../http/http';

export class CreateFolder201Response {
    'status'?: string;
    'data'?: CreateFolder201ResponseData;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "status",
            "baseName": "status",
            "type": "string",
            "format": ""
        },
        {
            "name": "data",
            "baseName": "data",
            "type": "CreateFolder201ResponseData",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return CreateFolder201Response.attributeTypeMap;
    }

    public constructor() {
    }
}