// object-storage/models/CreateFolder201ResponseData.ts
/**
 * Object Storage API Documentaion
 * A scalable, reliable, and cost effective Object Storage solution to support your application from liara-cloud
 *
 * OpenAPI spec version: 1.0.0
 * Contact: support@liara.ir
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class CreateFolder201ResponseData {
    'folder'?: string;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "folder",
            "baseName": "folder",
            "type": "string",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return CreateFolder201ResponseData.attributeTypeMap;
    }

    public constructor() {
    }
}